replace some calls to unpack with nativecast
via NativeCall - we can use NativeCall to get closer to the metal
and thus get our data types as we want them. this eliminates the
need for most of the calls to pack/unpack and thus a) completes
the readFloat routine[1] and b) speeds up lookup somewhat

sources:

    * http://www.nntp.perl.org/group/perl.perl6.users/2016/04/msg2716.html
    * https://www.mail-archive.com/search?l=perl6-all@perl.org&q=subject:%22Re%5C%3A+Blobs+and+IEEE+floating+point%22&o=newest&f=1
    * https://docs.perl6.org/language/nativecall#sub_nativecast

so now the module is passing it tests. next step is to add more
tests and make the code more idiomatic

[1] note, not tested on a big-endian system