remove some more redundant methods
get_database_version and get_module_version we neither care about
or use. remove 'return' when it is the last statement in a method
as it's more idiomatic to not use that. remove INVALID_IP_ADDRESS
as we can't *have* an invalid ip address anymore due to the IPv4
type constraint